tons of bugfixes, now runs on edison using 12 voices, midi is still lagging heavily preconfigured midi for the mpk25

+++ edisonSynth_linux/src/Oscillator.h
@@ -34,7 +34,7 @@ public:
 	Oscillator();
 	Oscillator(short*** wavetable);
 	short get_nextval();
-	short compute_nextval();
+	double compute_nextval();
 	int get_waveform();
 	void set_waveform(int wf);
 	void reset();
